\name{print.bdc}
\alias{print.bdc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print Method for Best Distance Correlation
}
\description{
Print the result of best distance correlation
}
\usage{
\method{print}{bdc}(bdc.result,...)
%print.bdc(bdc.result)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bdc.result}{a class of bdc}
  \item{...}{further arguments passed to or from other methods}
}
\value{
Print the summary of result of best distance correlation to stdout.
}
\references{
Cao X, Sahr N and Pounds S (2020)  Gene-Set Distance Associations (GSDA): A Powerful Tool for Gene-Set Association Analysis.
}
\author{
Xueyuan Cao <xcao12@uthsc.edu> and Stanley.pounds <stanley.pounds@stjude.org>
}
\seealso{
print.dcor print.gsda.result
}
