\name{write.gsda.csv.file}
\alias{write.gsda.csv.file}
\title{
Write GSDA Result to a Comma Delimited File
}
\description{
Write a gene-set distance analysis result to a comma delimited file (.csv)
}
\usage{
write.gsda.csv.file(gsda.result, out.file)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{gsda.result}{A class of gene-set distance analysis result}
  \item{out.file}{A .csv file name with directory}
}
\value{
A saved .csv file.
}
\author{
Xueyuan Cao <xcao12@uthsc.edu> and Stanley.pounds <stanley.pounds@stjude.org>
}
