\name{print.gsda.result}
\alias{print.gsda.result}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print Method for GSDA Result
}
\description{
Print the of result of gene-set distance analysis.
}
\usage{
\method{print}{gsda.result}(gsda.result)
%print.gsda.result(gsda.result)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{gsda.result}{result of gene-set distance analysis}
}
\details{
Print the result of gene-set distance analysis to stdout.
}
\references{
Cao X, Sahr N and Pounds S (2020)  Gene-Set Distance Associations (GSDA): A Powerful Tool for Gene-Set Association Analysis.
}
\author{
Xueyuan Cao <xcao12@uthsc.edu> and Stanley.pounds <stanley.pounds@stjude.org>
}
\seealso{
print.dcor print.gsda.result
}
