\name{print.gsda.result}
\alias{print.gsda.result}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print Method for GSDA Result
}
\description{
Print the of result of gene-set distance analysis (gsda function).
}
\usage{
\method{print}{gsda.result}(gsda.result)
%print.gsda.result(gsda.result)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{gsda.result}{result of gene-set distance analysis (gsda function)}
}
\details{
Print the result of gene-set distance analysis to stdout.
}
\references{
Cao X and Pounds S (2020)  Gene-Set Distance Associations (GSDA): A Powerful Tool for Gene-Set Association Analysis.
}
\author{Xueyuan Cao \email{xcao12@uthsc.edu} and Stanley Pounds \email{stanley.pounds@stjude.org}
}
\seealso{\code{\link{print.dcor}}, \code{\link{print.gsda.result}}}
